[#41047] Logger: Change LogUnhandledWrite argument type

+++ src/Emulator/Cores/Arm/Timers/CortexAGenericTimer.cs
@@ -1,5 +1,5 @@
 //
-// Copyright (c) 2010-2022 Antmicro
+// Copyright (c) 2010-2023 Antmicro
 // Copyright (c) 2011-2015 Realtime Embedded
 //
 // This file is licensed under the MIT License.
@@ -97,7 +97,7 @@ namespace Antmicro.Renode.Peripherals.Timers
                 }
                 break;
             default:
-                this.LogUnhandledWrite(offset, (long)value);
+                this.LogUnhandledWrite(offset, value);
                 break;
             }
         }
